Moved locking code to separate DoorManager code
