+++ ForgetMeNot/Assets/DoorsManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3d25e534dd06b846816f1bcb292d3a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
